+++ changelog.md
@@ -1,9 +1,10 @@
 # dspic33a-crc-basic v1.0.0
+
 ### Release Highlights
 
-
+- Initial release
 
 ### Features Added\Updated
 
-
-
+- The code example showcases CRC computation using the MCC Melody CRC Driver APIs.
+- Compares CRC-16-CCITT hardware CRC computation to software CRC computation.